requirements.txt
fastapi
uvicorn[standard]
sse-starlette
openai
biopython
alphagenome
pydantic
requests
python-dotenv

